'Refactored by Sourcery'

+++ userbot/plugins/locks.py
@@ -18,9 +18,7 @@ async def _(event):
     peer_id = event.chat_id
     if input_str in (("bots", "commands", "email", "forward", "url")):
         update_lock(peer_id, input_str, True)
-        await event.edit(
-            "Locked {}".format(input_str)
-        )
+        await event.edit(f"Locked {input_str}")
     else:
         msg = None
         media = None
@@ -90,9 +88,7 @@ async def _(event):
     peer_id = event.chat_id
     if input_str in (("bots", "commands", "email", "forward", "url")):
         update_lock(peer_id, input_str, False)
-        await event.edit(
-            "UnLocked {}".format(input_str)
-        )
+        await event.edit(f"UnLocked {input_str}")
     else:
         await event.edit(
             "Use `.lock` without any parameters to unlock API locks"
@@ -104,16 +100,18 @@ async def _(event):
     if event.fwd_from:
         return
     res = ""
-    current_db_locks = get_locks(event.chat_id)
-    if not current_db_locks:
-        res = "There are no DataBase locks in this chat"
-    else:
-        res = "Following are the DataBase locks in this chat: \n"
-        res += "👉 `bots`: `{}`\n".format(current_db_locks.bots)
+    if current_db_locks := get_locks(event.chat_id):
+        res = (
+            "Following are the DataBase locks in this chat: \n"
+            + "👉 `bots`: `{}`\n".format(current_db_locks.bots)
+        )
+
         res += "👉 `commands`: `{}`\n".format(current_db_locks.commands)
         res += "👉 `email`: `{}`\n".format(current_db_locks.email)
         res += "👉 `forward`: `{}`\n".format(current_db_locks.forward)
         res += "👉 `url`: `{}`\n".format(current_db_locks.url)
+    else:
+        res = "There are no DataBase locks in this chat"
     current_chat = await event.get_chat()
     try:
         current_api_locks = current_chat.default_banned_rights
@@ -140,9 +138,8 @@ async def check_incoming_messages(event):
     # TODO: exempt admins from locks
     peer_id = event.chat_id
     if is_locked(peer_id, "commands"):
-        entities = event.message.entities
         is_command = False
-        if entities:
+        if entities := event.message.entities:
             for entity in entities:
                 if isinstance(entity, types.MessageEntityBotCommand):
                     is_command = True
@@ -154,19 +151,17 @@ async def check_incoming_messages(event):
                     "I don't seem to have ADMIN permission here. \n`{}`".format(str(e))
                 )
                 update_lock(peer_id, "commands", False)
-    if is_locked(peer_id, "forward"):
-        if event.fwd_from:
-            try:
-                await event.delete()
-            except Exception as e:
-                await event.reply(
-                    "I don't seem to have ADMIN permission here. \n`{}`".format(str(e))
-                )
-                update_lock(peer_id, "forward", False)
+    if is_locked(peer_id, "forward") and event.fwd_from:
+        try:
+            await event.delete()
+        except Exception as e:
+            await event.reply(
+                "I don't seem to have ADMIN permission here. \n`{}`".format(str(e))
+            )
+            update_lock(peer_id, "forward", False)
     if is_locked(peer_id, "email"):
-        entities = event.message.entities
         is_email = False
-        if entities:
+        if entities := event.message.entities:
             for entity in entities:
                 if isinstance(entity, types.MessageEntityEmail):
                     is_email = True
@@ -179,9 +174,8 @@ async def check_incoming_messages(event):
                 )
                 update_lock(peer_id, "email", False)
     if is_locked(peer_id, "url"):
-        entities = event.message.entities
         is_url = False
-        if entities:
+        if entities := event.message.entities:
             for entity in entities:
                 if isinstance(entity, (types.MessageEntityTextUrl, types.MessageEntityUrl)):
                     is_url = True
@@ -199,34 +193,35 @@ async def check_incoming_messages(event):
 async def _(event):
     # TODO: exempt admins from locks
     # check for "lock" "bots"
-    if is_locked(event.chat_id, "bots"):
+    if not is_locked(event.chat_id, "bots"):
+        return
         # bots are limited Telegram accounts,
         # and cannot join by themselves
-        if event.user_added:
-            users_added_by = event.action_message.from_id
-            is_ban_able = False
-            rights = types.ChatBannedRights(
-                until_date=None,
-                view_messages=True
+    if event.user_added:
+        users_added_by = event.action_message.from_id
+        is_ban_able = False
+        rights = types.ChatBannedRights(
+            until_date=None,
+            view_messages=True
+        )
+        added_users = event.action_message.action.users
+        for user_id in added_users:
+            user_obj = await borg.get_entity(user_id)
+            if user_obj.bot:
+                is_ban_able = True
+                try:
+                    await borg(functions.channels.EditBannedRequest(
+                        event.chat_id,
+                        user_obj,
+                        rights
+                    ))
+                except Exception as e:
+                    await event.reply(
+                        "I don't seem to have ADMIN permission here. \n`{}`".format(str(e))
+                    )
+                    update_lock(event.chat_id, "bots", False)
+                    break
+        if Config.G_BAN_LOGGER_GROUP is not None and is_ban_able:
+            ban_reason_msg = await event.reply(
+                f"!warn [user](tg://user?id={users_added_by}) Please Do Not Add BOTs to this chat."
             )
-            added_users = event.action_message.action.users
-            for user_id in added_users:
-                user_obj = await borg.get_entity(user_id)
-                if user_obj.bot:
-                    is_ban_able = True
-                    try:
-                        await borg(functions.channels.EditBannedRequest(
-                            event.chat_id,
-                            user_obj,
-                            rights
-                        ))
-                    except Exception as e:
-                        await event.reply(
-                            "I don't seem to have ADMIN permission here. \n`{}`".format(str(e))
-                        )
-                        update_lock(event.chat_id, "bots", False)
-                        break
-            if Config.G_BAN_LOGGER_GROUP is not None and is_ban_able:
-                ban_reason_msg = await event.reply(
-                    "!warn [user](tg://user?id={}) Please Do Not Add BOTs to this chat.".format(users_added_by)
-                )
